set receiver cell color from power

+++ receiver.h
@@ -6,14 +6,14 @@
 class Receiver : public QGraphicsRectItem // each cell of the simulation grid acts like a receiver
 {
 public:
-    Receiver(int power_dBm);
+    Receiver(double power_dBm);
 
-    int getPower_dBm();
+    double getPower_dBm();
     qulonglong getBitrate();
     QColor getCellColor();
 
 private:
-    int power_dBm;
+    double power_dBm;
     qulonglong bitrate;
     QColor cell_color;
 };
